function/thumbnail-http/alibaba_thumbnail/requirements.txt
Pillow
boto3
oss2
ntplib
smlibrary